services/document-delivery-service/requirements.txt
Flask==2.0.2
Jinja2==3.0.3
MarkupSafe==2.0.1
PyNaCl==1.5.0
PyPDF2==1.26.0
Werkzeug==2.0.2
bcrypt==3.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
expiringdict==1.1.4
google-api-core==2.4.0
google-auth==2.3.3
google-cloud-core==2.2.1
google-cloud-storage==2.0.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
launchdarkly-server-sdk==7.3.0
paramiko==2.9.2
protobuf==3.19.3
pyRFC3339==1.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
semver==2.13.0
six==1.16.0
urllib3==1.26.8
git+https://github.com/daxiom/service-runner.py.git@0.0.6
git+https://github.com/daxiom/simple-cloudevent.py.git@0.0.2
